OBS url handling: use drop-in replacement for reuse of http(s) connections
The python osc module does not support reuse of http connection. This
means that we lose a lot of time just establishing connections. To avoid
changing the osc module for the moment we're using a terrible hack to
wrap the urlopen call and redirect it to httplib instead.

Performance will benefit quite from this change. Even if it ain't
pretty.