global: refactor cache check/refresh sequences throughout the code
A couple of statements in the update() functions have been repetitive
and it was easy to forget them which would break the cache. This is now
centralized and the update() functions can concentrate on actually
updating stuff. Except for more complex cases, where checking the cache
validity requires implementation specific logic.